Now have full shared server logic
