+++ src/Compilers/CSharp/Test/CommandLine/TouchedFileLoggingTests.cs
@@ -215,7 +215,7 @@ public class C { }").Path;
                 filelist.Add(source1);
                 var outWriter = new StringWriter();
                 var cmd = new CSharpCompilerServer(
-                    new DesktopCompilerServerHost(),
+                    new DesktopCompilerServerHost(Guid.NewGuid().ToString()),
                     new[] { "/nologo", "/touchedfiles:" + touchedBase, source1 },
                     null,
                     _baseDirectory,
