+++ src/Compilers/Server/ServerShared/DiagnosticListener.cs
