+++ src/Compilers/Server/ServerShared/IClientConnection.cs
