+++ src/Compilers/Server/ServerShared/ICompilerServerHost.cs
@@ -5,14 +5,16 @@ using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Linq;
 using System.Text;
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace Microsoft.CodeAnalysis.CompilerServer
 {
-    public interface ICompilerServerHost
+    internal interface ICompilerServerHost
     {
         IAnalyzerAssemblyLoader AnalyzerAssemblyLoader { get; }
         Func<string, MetadataReferenceProperties, PortableExecutableReference> AssemblyReferenceProvider { get; }
+        Task<IClientConnection> CreateListenTask(CancellationToken cancellationToken);
 
         string GetSdkDirectory();
         bool CheckAnalyzers(string baseDirectory, ImmutableArray<CommandLineAnalyzerReference> analyzers);
