+++ src/Compilers/Server/VBCSCompilerTests/CompilerServerApiTest.cs
@@ -148,6 +148,12 @@ class Hello
             return requestHandler;
         }
 
+        private static Mock<ICompilerServerHost> CreateNopCompilerServerHost()
+        {
+            var host = new Mock<ICompilerServerHost>();
+            return host;
+        }
+
         [Fact]
         public void NotifyCallBackOnRequestHandlerException()
         {
@@ -167,7 +173,7 @@ class Hello
                 Assert.Same(ex, providedEx);
                 invoked = true;
             });
-            var client = new ServerDispatcher.Connection(clientConnection, handler.Object);
+            var client = new Connection(clientConnection, handler.Object);
 
             Assert.Throws(typeof(AggregateException), () => client.ServeConnection().Wait());
             Assert.True(invoked);
@@ -194,7 +200,7 @@ class Hello
                 })
                 .Returns(s_emptyBuildResponse);
 
-            var client = new ServerDispatcher.Connection(clientConnection, handler.Object);
+            var client = new Connection(clientConnection, handler.Object);
             var serveTask = client.ServeConnection();
 
             // Once this returns we know the Connection object has kicked off a compilation and 
@@ -203,7 +209,7 @@ class Hello
             var cancellationToken = handlerTaskSource.Task.Result;
             monitorTaskSource.SetResult(true);
 
-            Assert.Equal(ServerDispatcher.CompletionReason.ClientDisconnect, serveTask.Result.CompletionReason);
+            Assert.Equal(CompletionReason.ClientDisconnect, serveTask.Result.CompletionReason);
             Assert.True(cancellationToken.IsCancellationRequested);
 
             // Now that the asserts are done unblock the "build" long running task.  Have to do this
@@ -221,8 +227,8 @@ class Hello
             clientConnection.ReadBuildRequestTask = TaskFromException<BuildRequest>(ex);
             clientConnection.CloseAction = delegate { calledClose = true; };
 
-            var client = new ServerDispatcher.Connection(clientConnection, handler.Object);
-            Assert.Equal(ServerDispatcher.CompletionReason.CompilationNotStarted, client.ServeConnection().Result.CompletionReason);
+            var client = new Connection(clientConnection, handler.Object);
+            Assert.Equal(CompletionReason.CompilationNotStarted, client.ServeConnection().Result.CompletionReason);
             Assert.True(calledClose);
         }
 
@@ -242,19 +248,18 @@ class Hello
                 .Setup(x => x.HandleRequest(It.IsAny<BuildRequest>(), It.IsAny<CancellationToken>()))
                 .Returns(s_emptyBuildResponse);
 
-            var client = new ServerDispatcher.Connection(clientConnection, handler.Object);
-            Assert.Equal(ServerDispatcher.CompletionReason.ClientDisconnect, client.ServeConnection().Result.CompletionReason);
+            var client = new Connection(clientConnection, handler.Object);
+            Assert.Equal(CompletionReason.ClientDisconnect, client.ServeConnection().Result.CompletionReason);
         }
 
         [Fact]
         public void KeepAliveNoConnections()
         {
             var keepAlive = TimeSpan.FromSeconds(3);
-            var pipeName = Guid.NewGuid().ToString();
             var requestHandler = new Mock<IRequestHandler>(MockBehavior.Strict);
-            var dispatcher = new ServerDispatcher(requestHandler.Object, new EmptyDiagnosticListener());
+            var dispatcher = new ServerDispatcher(CreateNopCompilerServerHost().Object, requestHandler.Object, new EmptyDiagnosticListener());
             var startTime = DateTime.Now;
-            dispatcher.ListenAndDispatchConnections(pipeName, keepAlive);
+            dispatcher.ListenAndDispatchConnections(keepAlive);
 
             Assert.True((DateTime.Now - startTime) > keepAlive);
         }
@@ -262,13 +267,13 @@ class Hello
         [Fact]
         public async Task FailedConnectionShouldCreateFailedConnectionData()
         {
-            var tcs = new TaskCompletionSource<NamedPipeServerStream>();
+            var tcs = new TaskCompletionSource<IClientConnection>();
             var handler = new Mock<IRequestHandler>(MockBehavior.Strict);
             var connectionDataTask = ServerDispatcher.CreateHandleConnectionTask(tcs.Task, handler.Object, CancellationToken.None);
 
             tcs.SetException(new Exception());
             var connectionData = await connectionDataTask.ConfigureAwait(false);
-            Assert.Equal(ServerDispatcher.CompletionReason.CompilationNotStarted, connectionData.CompletionReason);
+            Assert.Equal(CompletionReason.CompilationNotStarted, connectionData.CompletionReason);
             Assert.Null(connectionData.KeepAlive);
         }
 
@@ -283,8 +288,8 @@ class Hello
             var pipeName = Guid.NewGuid().ToString();
             var dispatcherTask = Task.Run(() =>
             {
-                var dispatcher = new ServerDispatcher(CreateNopRequestHandler().Object, listener);
-                dispatcher.ListenAndDispatchConnections(pipeName, keepAlive);
+                var dispatcher = new ServerDispatcher(CreateNopCompilerServerHost().Object, CreateNopRequestHandler().Object, listener);
+                dispatcher.ListenAndDispatchConnections(keepAlive);
             });
 
             await RunCSharpCompile(pipeName, HelloWorldSourceText).ConfigureAwait(false);
@@ -306,8 +311,11 @@ class Hello
             var pipeName = Guid.NewGuid().ToString();
             var dispatcherTask = Task.Run(() =>
             {
-                var dispatcher = new ServerDispatcher(new CompilerRequestHandler(Temp.CreateDirectory().Path), listener);
-                dispatcher.ListenAndDispatchConnections(pipeName, keepAlive);
+                var dispatcher = new ServerDispatcher(
+                    CreateNopCompilerServerHost().Object,
+                    new CompilerRequestHandler(CreateNopCompilerServerHost().Object, Temp.CreateDirectory().Path), 
+                    listener);
+                dispatcher.ListenAndDispatchConnections(keepAlive);
             });
 
             for (int i = 0; i < 5; i++)
@@ -332,8 +340,11 @@ class Hello
             var pipeName = Guid.NewGuid().ToString();
             var dispatcherTask = Task.Run(() =>
             {
-                var dispatcher = new ServerDispatcher(new CompilerRequestHandler(Temp.CreateDirectory().Path), listener);
-                dispatcher.ListenAndDispatchConnections(pipeName, keepAlive);
+                var dispatcher = new ServerDispatcher(
+                    CreateNopCompilerServerHost().Object,
+                    new CompilerRequestHandler(CreateNopCompilerServerHost().Object, Temp.CreateDirectory().Path), 
+                    listener);
+                dispatcher.ListenAndDispatchConnections(keepAlive);
             });
 
             var list = new List<Task>();
@@ -372,8 +383,8 @@ class Hello
             var pipeName = Guid.NewGuid().ToString();
             var dispatcherTask = Task.Run(() =>
             {
-                var dispatcher = new ServerDispatcher(CreateNopRequestHandler().Object, diagnosticListener.Object);
-                dispatcher.ListenAndDispatchConnections(pipeName, TimeSpan.FromSeconds(1), cancellationToken: cts.Token);
+                var dispatcher = new ServerDispatcher(CreateNopCompilerServerHost().Object, CreateNopRequestHandler().Object, diagnosticListener.Object);
+                dispatcher.ListenAndDispatchConnections(TimeSpan.FromSeconds(1), cancellationToken: cts.Token);
             });
 
             var seconds = 10;
